LayoutTests/fast/dom/HTMLDocument/named-item-multiple-match.html

<!DOCTYPE html>
<html>
<body>
<p>
Tests that the named item created for an image with an ID is correctly removed.
The test passes, if you see a "PASS" message in the div below.
</p>
<form name="someName" title="first"></form>
<form name="someName" title="second"></form>
<script src="../../js/resources/js-test-pre.js"></script>
<script>

var collection;
shouldBe('collection = window.someName; collection.length', '2');
shouldBe('collection[0].title', '"first"');
shouldBe('collection[1].title', '"second"');
shouldBe('collection[1]; window.someName', 'collection');

shouldBe('collection = document.someName; collection.length', '2');
shouldBe('collection[0].title', '"first"');
shouldBe('collection[1].title', '"second"');
shouldBe('collection[1]; document.someName', 'collection');

</script>
</body>
</html>
