third_party/blink/web_tests/platform/generic/fast/dom/HTMLDocument/named-item-multiple-match-expected.txt

PASS collection = window.someName; collection.length is 2
PASS collection[0].title is "first"
PASS collection[1].title is "second"
PASS collection[1]; window.someName is collection
PASS collection = document.someName; collection.length is 2
PASS collection[0].title is "first"
PASS collection[1].title is "second"
PASS collection[1]; document.someName is collection
PASS successfullyParsed is true

TEST COMPLETE


Tests that the named item created for an image with an ID is correctly removed. The test passes, if you see a "PASS" message in the div below.
